msi: extract arch-agnostic logic into a common file
In preparation to resolve #1088, this patch extracts
the architecture agnostic logic from arch/x64/msi.cc into a
new common file - drivers/msi.cc. The arch-specific functionality
comprising of the msix_vector::set_affinity() and some interrupt_manager
methods stay in the arch/<arch>/msi.cc files.

In addition, this patch changes the signature of the msix_vector::set_affinity()
method by replacing the x86_64 specific apic_id with the generic
sched::cpu() pointer.

Refers #1088
