+++ arch/aarch64/exceptions.hh
@@ -50,6 +50,9 @@ public:
     void register_interrupt(interrupt *interrupt);
     void unregister_interrupt(interrupt *interrupt);
 
+    unsigned register_handler(std::function<void ()> handler);
+    void unregister_handler(unsigned vector);
+
     /* invoke_interrupt returns false if unhandled */
     bool invoke_interrupt(unsigned int id);
